requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-jwt>=1.11,<2.0
psycopg2-binary>=2.8,<3.0
drf-yasg>=1.20,<2.0
gunicorn>=20.0,<21.0
drf_spectacular